// src/features/classes/model/classTypes.ts
export interface ClassItem {
    id: number;
    subjectId: number;
    subjectName: string;
    gradeId: number;
    gradeCode: string;
    gradeName: string;
    locationId: number;
    locationName: string;
    locationCity: string;
    durationId: number;
    durationName: string;
    durationMinutes: number;
    fee: number;
    requirements: string;
    note: string;
    studentGender: string;
    studentDescription: string;
    createdById: number;
    createdByName: string
}